requirements.txt
aiomysql==0.2.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
Werkzeug==3.0.4
